Remove extraneous toLong

+++ sql/core/src/main/scala/org/apache/spark/sql/execution/datasources/PruneFileSourcePartitions.scala
@@ -86,6 +86,6 @@ private[sql] object PruneFileSourcePartitions extends Rule[LogicalPlan] {
     } else {
       1.0
     }
-    (sizeInBytes.toLong * factor).toLong
+    (sizeInBytes * factor).toLong
   }
 }
